requirement.txt
adafruit-ampy==1.1.0
altgraph==0.17
asgiref==3.2.10
attrs==20.1.0
auto-py-to-exe==2.7.8
backcall==0.2.0
bitstring==3.1.7
bottle==0.12.18
bottle-websocket==0.2.9
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
decorator==4.4.2
Django==3.1
django-environ==0.4.5
django-extensions==3.0.6
django-filter==2.4.0
django-seed==0.2.2
djangorestframework==3.12.4
ecdsa==0.16.1
Eel==0.12.4
esptool==3.0
Faker==4.1.2
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
importlib-metadata==1.7.0
iniconfig==1.0.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.3.4
more-itertools==8.5.0
mysqlclient==2.0.1
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.7
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pyparsing==2.4.7
pyserial==3.5
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2020.1
reedsolo==1.5.4
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
wcwidth==0.2.5
whichcraft==0.6.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
